% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadExamples.R
\name{loadExamples}
\alias{loadExamples}
\title{R2RESUME: Converts Rmarkdown to Beautiful Resume Sites}
\usage{
loadExamples(template = "manzana", ...)
}
\arguments{
\item{template}{Default is the manzana, but you may declare a different template name
for the final look of your resume site. The options include:

manzana, limon, circuela, anana, naranja}
}
\description{
See the \href{https://www.r2resume.com}{Official R2RESUME
website} for additional details on using the \code{generate}
function.
}
\details{
Load an example starter resume template
}
\section{Brief}{


 When called, a new folder is created in your current directory and the requested example
 is copied into this folder. Finally, a call is automatically made to render the template
 and create a sample resume site.

 To edit and further build the template, you may proceed into the directory and manipulate
 the .Rmd files. You may also create other .Rmd files and it will be rendered simultaneously
}

\examples{
Example of calling this function within your document
\dontrun{

rmarkdown::loadExamples('naranja')


}
}
