% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPublication.R
\name{getPublication}
\alias{getPublication}
\title{R2RESUME: Converts Rmarkdown to Beautiful Resume Sites}
\usage{
getPublication(PMID, ...)
}
\arguments{
\item{PMID}{Required. Enter one or multiple PMIDs.Ensure to concatenate the PMIDs.
See examples below

PMID stands for pubmed ID. You may obtain this directly from www.ncbi.nlm.nih.gov/pubmed beneath your article.
In case you have many articles, you can quickly get a list of your PMID by selecting format to PMID.
Should you need assistance with this, shoot me a quick email at idonshayo at gmail.com

Required
Rtools  https://cran.rstudio.com/bin/windows/Rtools}
}
\description{
See the \href{https://www.r2resume.com}{Official R2RESUME
website} for additional details on using the \code{generate}
function.
}
\details{
This function simply takes an input of one or multiple PMID(PUBMED ID)
}
\examples{
Example when calling this function within your document
\dontrun{

```{r echo =FALSE}
PMID <- c(31645732,31634902,31636404,31657530,31645761,31645758,31653761,31642025,31647023,31645452,31647500,27976715)
```
`r RProfilesPackage::getPublication(PMID)`

}
}
